+++ meta_package_manager/tests/test_cli_search.py
@@ -91,26 +91,26 @@ class TestSearch(CLISubCommandTests, CLITableTests):
 
     def test_exact_search_tokenizer(self, invoke):
         result = invoke(
-            '--manager', 'pip3', 'search', '--exact', 'sed', color=False)
+            '--manager', 'pip', 'search', '--exact', 'sed', color=False)
         assert result.exit_code == 0
         assert "1 package total" in result.stdout
         assert " sed " in result.stdout
 
         for query in ['SED', 'SeD', 'sEd*', '*sED*', '_seD-@', '', '_']:
-            result = invoke('--manager', 'pip3', 'search', '--exact', query)
+            result = invoke('--manager', 'pip', 'search', '--exact', query)
             assert result.exit_code == 0
             assert "0 package total" in result.stdout
             assert "sed" not in result.stdout
 
     def test_fuzzy_search_tokenizer(self, invoke):
         for query in ['', '_', '_seD-@']:
-            result = invoke('--manager', 'pip3', 'search', query)
+            result = invoke('--manager', 'pip', 'search', query)
             assert result.exit_code == 0
             assert "0 package total" in result.stdout
             assert "sed" not in result.stdout
 
         for query in ['sed', 'SED', 'SeD', 'sEd*', '*sED*']:
-            result = invoke('--manager', 'pip3', 'search', query, color=False)
+            result = invoke('--manager', 'pip', 'search', query, color=False)
             assert result.exit_code == 0
             assert "2 packages total" in result.stdout
             assert " sed " in result.stdout
@@ -118,12 +118,12 @@ class TestSearch(CLISubCommandTests, CLITableTests):
 
     def test_extended_search_tokenizer(self, invoke):
         for query in ['', '_', '_seD-@']:
-            result = invoke('--manager', 'pip3', 'search', '--extended', query)
+            result = invoke('--manager', 'pip', 'search', '--extended', query)
             assert result.exit_code == 0
             assert "0 package total" in result.stdout
             assert "sed" not in result.stdout
 
         for query in ['sed', 'SED', 'SeD', 'sEd*', '*sED*']:
-            result = invoke('--manager', 'pip3', 'search', '--extended', query)
+            result = invoke('--manager', 'pip', 'search', '--extended', query)
             assert result.exit_code == 0
             assert "23 packages total" in result.stdout