+++ meta_package_manager/tests/test_docs.py
@@ -45,4 +45,4 @@ def test_changelog():
             entry = match.groupdict()
             assert entry['category']
             assert set(entry['category'].split(',')).issubset(flatten([
-                MANAGER_IDS, OS_DEFINITIONS.keys(), 'pip', 'mpm', 'bitbar']))
+                MANAGER_IDS, OS_DEFINITIONS.keys(), 'mpm', 'bitbar']))